+++ tests/testit/test-utils.R
@@ -28,15 +28,6 @@ assert("last author is prepend with and ", {
   (suppressWarnings(post_process_authors(unmodified)) %==% unmodified)
 })
 
-assert("all journals are listed and have a template folder", {
-  all <- grep("_article$", getNamespaceExports("CUPrticles"), value = TRUE)
-  all <- gsub("_article$", "", all)
-  folder_name <- journals()
-  # Special case for format function handling several version of template
-  folder_name <- unique(gsub("^([^_]+)_.*$", "\\1", folder_name))
-  (folder_name %==% sort(all))
-})
-
 assert("Named vector is transformed to pandoc variable args", {
   (vec_to_pandoc_variable_args(c(a = "b")) %==%
     c(rmarkdown::pandoc_variable_arg("a", "b"))